+++ backend/.gitignore
@@ -1,0 +1,36 @@
+# Ignore node_modules directory
+node_modules/
+
+# Ignore environment variables and sensitive files
+.env
+.env.local
+.env.*.local
+
+# Ignore build outputs
+dist/
+build/
+
+# Ignore logs
+logs/
+*.log
+npm-debug.log*
+
+# Ignore system files
+.DS_Store
+Thumbs.db
+
+# Ignore IDE configurations
+.vscode/
+.idea/
+*.iml
+
+# Ignore temporary files
+*.tmp
+*.swp
+*.bak
+
+# Ignore coverage reports
+coverage/
+
+# Ignore custom debug files
+debug.log
