+++ frontend/src/components/FAQ.jsx
