+++ frontend/src/components/Hero.jsx
